feat: comment out example code, but still test it in CI
We comment out many of the code blocks that would be
superfluous to add to every new plugin. However, we
comment them using a special '### ' prefix:

     # Normal comment
     ### this_is_example_code()
     ### another_line = of_example_code()

When we generate a plugin and run tests on it
for this repo, we remove the '### ' prefixes,
essentially enabling all example code for
testing purposes.

The '### ' prefix was chosen because it is relatively
unobtrusive to the plugin reader (who probably not
be aware of its significance) while still being
easy replace with sed. As a bonus, it's 4 characters
long, which plays nice with IDEs that indicate
indent levels.